+++ src/web-app/screens/Network/ActionsMenu.tsx
@@ -94,7 +94,7 @@ export const ActionsMenu: React.FC<ActionsMenuProps> = ({ network, withoutCreate
     [performActionCommand],
   );
   const startButton = withoutCreate ? null : (
-    <Button small intent={Intent.SUCCESS} text={t("Create")} icon={IconNames.PLUS} onClick={onCreateClick} />
+    <Button size="small" intent={Intent.SUCCESS} text={t("Create")} icon={IconNames.PLUS} onClick={onCreateClick} />
   );
   const removeWidget = network ? (
     <ConfirmMenu onConfirm={onRemove} tag={network.name} disabled={disabledAction === "network.remove"}></ConfirmMenu>
@@ -107,8 +107,8 @@ export const ActionsMenu: React.FC<ActionsMenuProps> = ({ network, withoutCreate
           <>
             {startButton ? <Divider /> : null}
             <Button
-              small
-              minimal
+              size="small"
+              variant="minimal"
               intent={Intent.NONE}
               title={t("Reload current list")}
               icon={IconNames.REFRESH}