+++ src/web-app/screens/Settings/Connection/EngineSelect.tsx
@@ -91,10 +91,10 @@ export const EngineSelect: React.FC<EngineSelectProps> = ({
         activeItem={activeItem}
       >
         <Button
-          alignText={Alignment.LEFT}
+          alignText={Alignment.START}
           disabled={disabled}
           fill
-          rightIcon={IconNames.CARET_DOWN}
+          endIcon={IconNames.CARET_DOWN}
           text={activeItem?.label ?? t("-- Select --")}
           textClassName={classNames({
             [Classes.TEXT_MUTED]: activeItem === undefined,
@@ -104,8 +104,8 @@ export const EngineSelect: React.FC<EngineSelectProps> = ({
       {withoutDetect ? null : (
         <>
           <Divider />
-          <ButtonGroup minimal>
-            <Button disabled={pending} small text={t("Detect")} intent={Intent.SUCCESS} onClick={onItemDetect} />
+          <ButtonGroup variant="minimal">
+            <Button disabled={pending} size="small" text={t("Detect")} intent={Intent.SUCCESS} onClick={onItemDetect} />
           </ButtonGroup>
         </>
       )}